pdf-transcriber/requirements.txt
pdf2image==1.17.0
pytesseract==0.3.13
opencv-python==4.13.0.92
python-dotenv==1.2.2
pillow==12.2.0
imutils==0.5.4
scipy==1.17.1
